requirements.txt
dnspython==2.0.0
emoji==0.6.0
emoji-country-flag==1.2.3
from==0.1.0
Pillow==8.0.1
pprint==0.1
pymongo==3.11.2
